requirements.txt
numpy
pillow
torch
torchvision
scikit-learn